Microdroid Mananger main.rs refactoring
I'm just re-ordering functions here, to try to make the startup
sequence easier to see. The various helper functions are moved down,
and the functions that do the actual work are moved up and in the
right order.

I also move one function only used in verify.rs to that file.

Bug: 308759880
Test: atest MicrodroidTests
Change-Id: I563036b4a23970b23f80598030d9e42a4b0e3425