Treat APK path change as compatible
The path to an APK may change, on an update, or at any other time when
it's not running. So we shouldn't treat a path change as rendering the
VM config incompatible.

(If the new path points to an APK with a different signature or
incompatible version then starting the VM may still fail - but we
can't determine that statically from the config, any more than we can
from checking package names.)

Modify the relevant test to match.

Bug: 321184334
Test: atest MicrodroidTests
Change-Id: I384360b0d19013b12da338af95cbc5ae104fa983