[attestation] Rename test_apk to demo_apk for Vm attestation
A separate test_apk will be added in a subsequent cl for e2e test.

Specifically, the demo_apk calls the real request attestation API
and interacts with RKPD to retrieve the real remotely provisioned
keys; while the test_apk will call the request attestation API
for testing only and will not trigger RKPD, a mock key will be used
in this case.

Test: Run VmAttestationDemoApp manually
Change-Id: I5e02dc071d167156e98088829227e83300899461