Use release_avf_microdroid_kernel_version to pick Microdroid kernel
This patch just sets up the directory structure for Microdroid kernel
prebuilts, and does the plumbing in Android.bp

Follow up patches will drop a android15-6.1 microdroid kernel prebuilt
into repo, and switch the trunk_staging to use it.

Once the android15-6.6 is stable to run as guest, we will switch to
that.

Bug: 298011555
Test: builds
Change-Id: I4f56e6c98e24709cf8461bded45b2c8035f48130