pvmfw: Parse cpu-map DT node for the guest
Support a subset of all possible topologies[*] where the guest sees a
set of clusters of cores with the usual read-then-patch approach. Note
that no separate validation is required as parsing ensures that the
nodes within a cluster properly start from 0 and are contiguous (same
for clusters) and that they point to valid CPU nodes. Extra properties,
if any are ignored.

Add labels to the cpu nodes so that DTC generates phandles for them,
which pvmfw can then use when patching the cpu-map node in the template
DT.

[*]: https://www.kernel.org/doc/Documentation/devicetree/bindings/cpu/cpu-topology.txt

Bug: 284369518
Test: m pvmfw
Change-Id: Ia70a73545b1740a3343ce7f0e4549da456cd6d55