pvmfw: Avoid panic if OPP table too large
As ArrayVec::push() panics if called on a full instance, ensure that it
is only called up to the capacity of the array and print a warning if
more entries are received.

Functionally, this will boot with fewer OPPs than actually supported
i.e. a potential performance impact but which is preferable to
preventing guests from booting altogether.

Bug: 335130218
Test: m pvmfw && TH
Change-Id: I7e064d87202170e670f3855c4c949b68c6429610