pvmfw: Leave UART shared for debuggable pVMs
When the payload is debuggable (detected through its VBMeta), pvmfw
allows any cmdline argument to be passed to the guest kernel. For Linux,
this means the host could enable earlycon through the 16550-inspired
UARTs. However, as the Microdroid kernel does not support MMIO_GUARD for
earlycon, this would not work (possibly resulting in a guest kernel
panic due to the unhandled injected exception? TBD).

To support this use-case, make pvmfw leave the UART mapped for the next
stage if it detects that the payload is debuggable (should not be used
in production & reflected in the payload's DICE chain). Note that
debuggable payloads that already support sharing the UART (which, from
the perspective of the hypervisor, would now be sharing it "twice")
should not break as the HVC still succeeds when called on already
guarded MMIO pages.

Test: m pvmfw
Bug: 356622725
Change-Id: I3f8cc3317142ec55d2c864bd658d42fa2c0ae345