vmbase: Remove dtb_region from linker scripts
Get rid of the region as it was never needed (our code does not generate
symbols for it), duplicates the AArch64 Linux boot protocol (which
states that the DT should be discovered through x0) used by AVF, and is
not even used by any production code. More importantly, it has caused
undesired confusion when debugging unrelated issues, for no added value
given that the testing coverage it provides is limited to itself.

Bug: 355696053
Test: m pvmfw rialto_unsigned vmbase_example_bin
Test: TH
Change-Id: I554ba106305ef5c84f20131ef377e5a46ca3d0ab