vmbase_example: Refactor DT boot pointer handling
Flip the logic so that, instead of using the linker-defined region, we
use the pointer passed through the boot protocol. The assertion ensures
that this won't result in a change in behavior but removes one
dependency on the linker definition.

To prepare for vmbase_example to be boot protocol compliant, pass the DT
region to init_page_table() instead of using a hard-coded assumption
about the DT region, which removes another use of the linker definition.

Bug: 355696053
Test: m pvmfw rialto_unsigned vmbase_example_bin
Test: TH
Change-Id: I7aa01d95139c492e76386aa737ce4afd54fe6779