vmbase_example: Clarify that PageTable is dropped
The code calls init_page_table() at some point and, based on the name of
the function and when it gets called (right before vmbase_example
accesses the PCI BARs), it seems to be to configure the page tables in
use. However, the PageTable struct is local to the function and gets
dropped when it returns. As a result, TTBR0_EL1 (which was reprogrammed
by page_table.activate()) reverts back to its previous value, which was
the hard-coded page tables in idmap.S!

The reason why the accesses to the PCI BAR or (right before) to the DT
don't result in page faults is because they respectively fall in the
DEVICE_REGION ([0; 1Gib[) and dtb_region, both of which are already
mapped in idmap.S.

Instead, keep the code unmodified but explicitly show that we're
dropping the page_table struct and mention it in a comment.

No functional change intended.

Test: atest vmbase_example.integration_test
Change-Id: Iaa2b21f42dc9afcebe2192fe7fdcfc4df80d01e8