Remove usages of select on defaults property
In order to support changing the global configuration per-module,
we can't make decisions on configuration until the configuration is
decided. The defaults mutator is one of the earliest mutators, and
it would be helpful to run it before deciding the configuration.

Remove usages of selects on defaults modules. Note that soong config
modules have the same problem, but they will just silently fall back
to using the true global configuration instead of throwing an error
that they tried to read the configuration before it was decided.
The usages of soong config variables on selects will have to be removed
in order to acheive a truely configuration-independant build.

Bug: 361816274
Test: Presubmits
Change-Id: I797f828d486081011cd71f7649544ef1bb326494