MicrodroidConfig: Unify API across flag configs
Flags should control the behavior of the code but using different APIs
with different flag configurations makes the code harder to reason about
when making a modification (in particular because flags configurations
aren't all tested at the same rate and a build breakage might be hard to
detect) so unfiy the different flag-defined "variants" of a same method
under a same signature.

In this process, improve the existing signatures:

 - prefer "-> Option<&T>" over "-> &Option<T>" as it's easier to
   generate, manipulate, and reason about

 - prefer "-> &[T]" over "-> &Vec<T>" as it is more flexible, hiding an
   implementation detail about the internal representation, and is
   typically the recommended type to use in an API

Note: No functional change intended.

Test: banchan com.android.virt aosp_arm64 && m apps_only dist
Change-Id: I240d0bb0f1d6c630c89f2939261eed021912d01d